Refactor auction prices to decimal, add DTOs & controller
- Changed Auction price fields from int to decimal for accuracy
- Added EF migration to update DB schema to numeric types
- Introduced AuctionDto, CreateAuctionDto, UpdateAuctionDto
- Added MappingProfile for AutoMapper DTO/entity mapping
- Implemented AuctionsController with CRUD endpoints
- Registered AutoMapper and switched to controller routing
- Added .editorconfig for C# code style and formatting
- Cleaned up .csproj Controllers folder entry

+++ src/Monart.AuctionService/Entities/Auction.cs
@@ -5,11 +5,11 @@ namespace Monart.AuctionService.Entities
     public class Auction
     {
         public Guid Id { get; set; }
-        public int ReservePrice { get; set; } = 0;
+        public decimal ReservePrice { get; set; } = 0;
         public string Seller { get; set; }
         public string Winner { get; set; }
-        public int? SoldAmount { get; set; }
-        public int? CurrentHighBid { get; set; }
+        public decimal? SoldAmount { get; set; }
+        public decimal? CurrentHighBid { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
         public DateTime AuctionEnd { get; set; }
